define 'DoHTTPGetRequest' as thingfulx.Client method
